requirements.txt
tqdm
typeguard
visualdl>=2.2.0
opencv-python
PyYAML>=5.1
shapely
scipy>=1.0.0
terminaltables
Cython
pycocotools
setuptools
prettytable
ujson
pillow
scikit-learn>=0.21.0
gast==0.3.3
faiss-cpu==1.7.1.post2
easydict
scikit-image
pyclipper
matplotlib
pandas
lap
motmetrics
sklearn
filterpy
bcolz
onnx
onnxruntime